Move Dev settings to the sidebar
