+++ web_src/main.js
@@ -20,81 +20,6 @@ const skipConfirms = browser.isCode;
 let map = null; // Leaflet map object containing current game map and all its markers
 
 //=================================================================================================
-// BuildMode handling
-
-export const buildMode = {
-  marker: undefined, // Current marker we're editing in Build Mode
-  object: undefined, // Current object we're editing in Build Mode
-  objectChanges: [], // Current object's changed values before they are committed to the list
-  changeList: [], // Changes made in the current Build Mode session
-};
-
-function toggleDevMode() {
-  Settings.globalSetDefault('devMode', false);
-  Settings.global.devMode = !Settings.global.devMode;
-  Settings.commit();
-  skipConfirms || alert('Dev mode is now set to ' + Settings.global.devMode + '.');
-}
-
-function toggleBuildMode() {
-  Settings.globalSetDefault('buildMode', false);
-  Settings.global.buildMode = !Settings.global.buildMode;
-  Settings.commit();
-  skipConfirms || alert('Build mode is now set to ' + Settings.global.buildMode + '.');
-}
-
-function updateBuildModeValue(event) {
-  let el = event.target;
-  let value = '{["'.includes((el.value + ' ').charAt(0)) ? JSON.parse(el.value) : el.value;
-  buildMode.object[el.id] = value;
-  buildMode.objectChanges[MapObject.makeAlt(buildMode.object.area, buildMode.object.name) + '|' + el.id] = value;
-}
-window.updateBuildModeValue = function (event) {
-  updateBuildModeValue(event);
-};
-
-function commitCurrentBuildModeChanges() {
-  Object.getOwnPropertyNames(buildMode.objectChanges).forEach(function (i) {
-    buildMode.changeList[i] = buildMode.objectChanges[i];
-  });
-  let newLat = buildMode.object.lat;
-  let newLng = buildMode.object.lng;
-
-  buildMode.marker.setLatLng(new L.LatLng(newLat, newLng));
-  buildMode.objectChanges = [];
-  map.closePopup();
-}
-window.commitCurrentBuildModeChanges = function () {
-  commitCurrentBuildModeChanges();
-};
-
-function exportBuildChanges() {
-  // It might be worth accummulating the changes in this structure as we make them, but this works
-  let jsonobj = {};
-  Object.getOwnPropertyNames(buildMode.changeList)
-    .filter(function (e) {
-      return e !== 'length';
-    })
-    .forEach(function (k) {
-      let alt, prop, area, name;
-      [alt, prop] = k.split('|');
-      [area, name] = alt.split(':');
-      if (!jsonobj[alt]) {
-        jsonobj[alt] = {};
-      }
-      jsonobj[alt]['name'] = name;
-      jsonobj[alt]['area'] = area;
-      jsonobj[alt][prop] = buildMode.changeList[k];
-    });
-  jsonobj = Object.values(jsonobj);
-
-  console.log(buildMode.changeList);
-  let t = JSON.stringify(jsonobj, null, 2);
-  browser.copyTextToClipboard(t);
-  skipConfirms || alert('Build mode changes have been placed on the clipboard.');
-}
-
-//=================================================================================================
 // setupKeyControls
 
 function setupKeyControls(map, searchControl) {
@@ -293,45 +218,6 @@ async function loadMap(mapParam) {
                 options: { toolbarIcon: { html: 'copy', tooltip: 'Copy pin positions to clip board' } },
                 addHooks: function () {
                   MapPins.copyToClipboard();
-                  subAction.prototype.addHooks.call(this); // closes sub-action
-                },
-              }),
-              subAction.extend({
-                options: { toolbarIcon: { html: '&times;', tooltip: 'Close' } },
-              }),
-            ],
-          }),
-        },
-      }),
-      // build mode button
-      L.Toolbar2.Action.extend({
-        options: {
-          toolbarIcon: { html: '<i class="fa-solid fa-screwdriver-wrench"></i>', tooltip: 'Developer Mode' },
-          subToolbar: new L.Toolbar2({
-            actions: [
-              subAction.extend({
-                options: { toolbarIcon: { html: 'Dev', tooltip: 'Toggles Developer mode on or off' } },
-                addHooks: function () {
-                  toggleDevMode();
-                  subAction.prototype.addHooks.call(this); // closes sub-action
-                },
-              }),
-              subAction.extend({
-                options: { toolbarIcon: { html: 'Build', tooltip: 'Toggles Build mode on or off' } },
-                addHooks: function () {
-                  toggleBuildMode();
-                  subAction.prototype.addHooks.call(this); // closes sub-action
-                },
-              }),
-              subAction.extend({
-                options: {
-                  toolbarIcon: {
-                    html: 'Copy Changes',
-                    tooltip: 'Copies the changes made in this session to the Clipboard',
-                  },
-                },
-                addHooks: function () {
-                  exportBuildChanges();
                   subAction.prototype.addHooks.call(this); // closes sub-action
                 },
               }),
